task appli v2

+++ src/app/task-list/task-list.component.ts
@@ -1,9 +1,11 @@
 import { Component, OnInit } from '@angular/core';
-import { Router } from '@angular/router';
-import { TaskService } from '../task-service.service';
+import { TaskService } from '../services/task-service.service';
 import { Task } from '../models/task';
 import { NgFor } from '@angular/common';
 import { MatListModule } from '@angular/material/list';
+import { RootingService } from '../services/rooting.service';
+import { MatDialog } from '@angular/material/dialog';
+import { AlertDialogComponent } from '../alert-dialog/alert-dialog.component';
 
 @Component({
   selector: 'app-task-list',
@@ -15,8 +17,9 @@ import { MatListModule } from '@angular/material/list';
 export class TaskListComponent implements OnInit {
 
   tasks: Task[];
-
-  constructor(private router: Router, private taskService: TaskService) { }
+  title = 'Title';
+  description = 'Description';
+  constructor(public readonly rootingService: RootingService, public readonly taskService: TaskService, private dialogPopUp: MatDialog) { }
 
   ngOnInit(): void {
     this.getRefreshTaskList();
@@ -29,16 +32,20 @@ export class TaskListComponent implements OnInit {
   }
 
   removeTask(task: Task) {
-    this.taskService.deleteTask(task.id!).subscribe({ next: () => {
-        this.getRefreshTaskList();}
+    this.dialogPopUp.open(AlertDialogComponent, {
+        data: {
+          message: 'Are you sure about deleting this task',
+          buttonback: 'No',
+          buttonDelete: 'Yes delete',
+        }
+      }).afterClosed()
+      .subscribe((confirmDeletingTask: Boolean) => {
+        if (confirmDeletingTask) {
+          this.taskService.deleteTask(task.id!).subscribe({ next: () => {
+            this.getRefreshTaskList();
+            }
+         });
+        }
     });
-  }
-
-  editTask(task: Task) {
-    this.router.navigate(['edit-task/' + task.id]);
-  }
-
-  addNewTask() {
-    this.router.navigate(['new-task'])
   }
 }
